requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
dblpy @ git+https://github.com/top-gg/python-sdk/@41a6b4486c175cfe87610907b44df4b31f484525
discord.py==1.7.3
firebase-admin==5.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.21.1
opencv-python-headless==4.5.3.56
packaging==21.0
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
ratelimiter==1.2.0.post0
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
yarl==1.6.3
